requirements_vision.txt
# CLIP Tile Recognition Dependencies
torch>=1.13.0
torchvision>=0.14.0
clip-by-openai==1.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.0  # For fast similarity search
requests>=2.28.0
base64
io